+++ frontend/package.json
@@ -23,6 +23,11 @@
         "tsparticles": "^3.8.1",
         "uuid": "^11.1.0",
         "web-vitals": "^5.0.0",
+        "workbox-core": "^7.3.0",
+        "workbox-expiration": "^7.3.0",
+        "workbox-precaching": "^7.3.0",
+        "workbox-routing": "^7.3.0",
+        "workbox-strategies": "^7.3.0",
         "zustand": "^5.0.4"
     },
     "scripts": {
