+++ frontend/src/serviceWorkerRegistration.d.ts
@@ -1,0 +1,2 @@
+export function register(config?: any): void;
+export function unregister(): void;
